Organize page posts flow, part 2. ref #33
