+++ src/common/StructureToArray.java
@@ -1,8 +1,8 @@
 package common;
 
 import common.structures.GroupStructure;
+import common.structures.PagePostsStructure;
 import common.structures.PageStructure;
-import main.PagePostsStructure;
 
 public class StructureToArray {
