+++ src/impl/FileXML.java
@@ -22,9 +22,9 @@ import common.Logger_;
 import common.StructureType;
 import common.structures.ConfigStructure;
 import common.structures.GroupStructure;
+import common.structures.PagePostsStructure;
 import common.structures.PageStructure;
-import main.PagePostsStructure;
-import main.SearchStructure;
+import common.structures.SearchStructure;
 
 public class FileXML {
