+++ src/main/Start.java
@@ -11,20 +11,22 @@ import common.Configurations;
 import common.Logger_;
 import common.structures.ConfigStructure;
 import common.structures.GroupStructure;
+import common.structures.PagePostsStructure;
 import common.structures.PageStructure;
+import common.structures.SearchStructure;
 import common.structures.ToPost;
 import impl.CheckNotifications;
 import impl.DoLogin;
 import impl.DoLogout;
 import impl.FileXML;
-import impl.PagePostsFiles;
-import impl.Posts;
 import impl.group.ManageFiles;
 import impl.group.Publish;
 import impl.groups.GoToGroups;
 import impl.groups.ManageGroupsListFiles;
 import impl.managePages.GoToManagePages;
 import impl.managePages.ManagePagesFiles;
+import impl.pagePosts.PagePostsFiles;
+import impl.pagePosts.Posts;
 
 public class Start {
